Compile using the -fvisibility=hidden flag c-icap and c-icap-library
This is a requirement for some linux distributions. 

Also it allow c-icap and c-icap library to export only symbols which
belongs to the c-icap and c-icap-lib APIs. There are many functions defined
for internal use and there is the risk that this functions/sympols be confused
with other functions/sympols defined in libraries and programs which linked
with c-icap/c-icap-lib.

This patch tested on Linux only. FreeBSD and Solaris should checked also.

This patch:
  - Enable -fvisibility=hidden flag when compiling c-icap and c-icap-lib
  - Add the '__attribute__ ((visibility ("default")))' gcc  to CI_DECLARE_DATA
    and CI_DECLARE_FUNC macros. This is will allow as to mark API functions
    as exported
  - Use the CI_DECLARE_FUNC/CI_DECLARE_DATA macros to many cases where must 
    used.
  - Fix ci_groups_ops to use only local/static defined functions. The
    stringdup/stringfree/stringlen used before defined in c-icap-lib and
    does not exported. The ci_groups_ops defined inside c-icap.
  - Move cfg_acl_add method from acl.cc file (included in c-icap-lib) to
    cfg_param.cc. Also implement the ci_acl_add_data API function to add
    acls, or data to acls

TODO:
   Check if "-fvisibility=hidden" works well for Solaris and FreeBSD using 
   g++ or clang